requirements.txt
flask==3.0.3
python-dotenv==1.0.1

langchain==0.3.7
langchain-core==0.3.17
langchain-community==0.3.7
langchain-text-splitters==0.3.2
langchain-pinecone==0.2.0
langchain-google-genai==2.0.1

pinecone==6.0.0

sentence-transformers==2.7.0
transformers==4.41.2
torch
pypdf

tqdm
numpy==1.26.4
pandas